% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaire-api.R
\name{api_params}
\alias{api_params}
\title{Parameters for the API}
\usage{
api_params(
  size = 10000,
  page = NULL,
  format = "xml",
  model = NULL,
  sort_by = NULL,
  sort_order = NULL,
  has_ec_funding = NULL,
  has_wc_funding = NULL,
  fp7_scientific_area = NULL,
  funder = NULL,
  funding_stream = NULL,
  proj_grant_id = NULL,
  proj_publication_id = NULL,
  proj_dataset_id = NULL,
  proj_title = NULL,
  proj_acronym = NULL,
  proj_call_id = NULL,
  proj_start_year = NULL,
  proj_end_year = NULL,
  proj_country = NULL,
  proj_org = NULL,
  r_sort_by = NULL,
  r_doi = NULL,
  r_orcid = NULL,
  r_from_date = NULL,
  r_to_date = NULL,
  r_title = NULL,
  r_author = NULL,
  r_provider_id = NULL,
  r_project_id = NULL,
  r_has_project = NULL,
  r_grant_id = NULL,
  r_fp7_grant_id = NULL,
  r_is_oa = NULL,
  r_country = NULL,
  pub_publication_id = NULL,
  pub_original_id = NULL,
  pub_sdg = NULL,
  pub_fos = NULL,
  dataset_id = NULL,
  software_id = NULL,
  other_id = NULL
)
}
\arguments{
\item{size}{by default 10000}

\item{page}{which page to return (when paging), default NULL}

\item{format}{the respone format, one of "xml", "json", "tsv", "csv", by default "xml"}

\item{model}{the model to use, by default NULL}

\item{sort_by}{sorting parameter}

\item{sort_order}{sorting order}

\item{has_ec_funding}{logical}

\item{has_wc_funding}{logical}

\item{fp7_scientific_area}{character}

\item{funder}{character}

\item{funding_stream}{character}

\item{proj_grant_id}{character}

\item{proj_publication_id}{character}

\item{proj_dataset_id}{character}

\item{proj_title}{character}

\item{proj_acronym}{character}

\item{proj_call_id}{character}

\item{proj_start_year}{character}

\item{proj_end_year}{character}

\item{proj_country}{character}

\item{proj_org}{character}

\item{r_sort_by}{character}

\item{r_doi}{character}

\item{r_orcid}{character}

\item{r_from_date}{character}

\item{r_to_date}{character}

\item{r_title}{character}

\item{r_author}{character}

\item{r_provider_id}{character}

\item{r_project_id}{character}

\item{r_has_project}{character}

\item{r_grant_id}{character}

\item{r_fp7_grant_id}{character}

\item{r_is_oa}{character}

\item{r_country}{character}

\item{pub_publication_id}{character}

\item{pub_original_id}{character}

\item{pub_sdg}{character}

\item{pub_fos}{character}

\item{dataset_id}{character}

\item{software_id}{character}

\item{other_id}{= NULL}
}
\value{
a list of parameters
}
\description{
Parameters for the API
}
