% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAIRE-package.R
\docType{package}
\name{OpenAIRE-package}
\alias{OpenAIRE-package}
\alias{OpenAIRE}
\title{\strong{R client for the OpenAIRE APIs}}
\description{
OpenAIRE is an open science initiative, which supports the
Open Access policy of the European Commission. This R package OpenAIRE
provides access from within R to the OpenAIRE RESTful APIs, which are
intended for metadata discovery and exploration only.
}
\author{
Markus Skyttner \email{markussk@kth.se}

Mohamad Bazzi \email{bazzi@kth.se}
}
