% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaire-api.R
\name{openaire_crawl}
\alias{openaire_crawl}
\title{Fetch several pages of results}
\usage{
openaire_crawl(
  entity = names(api_paths()),
  params = api_params(),
  page_size = 50L
)
}
\arguments{
\item{entity}{one of the different entity types}

\item{params}{a list of parameters}

\item{page_size}{page size, by default 50}
}
\value{
a data frame if possible
}
\description{
Fetch several pages of results
}
