% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaire-api.R
\name{openaire_projects}
\alias{openaire_projects}
\title{Projects given an acronym}
\usage{
openaire_projects(countries = "se", acronyms = "KTH")
}
\arguments{
\item{countries}{the countries to filter by, comma-separated string of 2-letter ISO country codes, by default "SE"}

\item{acronyms}{the acronyms to filter by, whitespace-separated string, by default "KTH"}
}
\description{
Projects given an acronym
}
