% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaire-api.R
\name{openaire}
\alias{openaire}
\title{Retrive OpenAIRE data}
\usage{
openaire(entity = names(api_paths()), page_size = 50, params = NULL)
}
\arguments{
\item{entity}{one of "projects", "research", "publications", "datasets", "software" or "other"}

\item{page_size}{records returned, by default 50}

\item{params}{a set of parameters to the API, see api_params()}
}
\value{
object with results, if possible a data frame
}
\description{
Retrive OpenAIRE data
}
